refactor: 캔버스 관련 컴포넌트 features와 widgets으로 분리
