+++ apps/frontend/src/App.tsx
@@ -2,10 +2,10 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
 import Sidebar from "./components/sidebar";
 import SideWrapper from "./components/layout/SideWrapper";
-import Canvas from "./components/canvas";
 
 import { useSyncedUsers } from "./hooks/useSyncedUsers";
 import { EditorView } from "./widgets/EditorView";
+import { CanvasView } from "./widgets/CanvasView";
 
 const queryClient = new QueryClient();
 
@@ -18,7 +18,7 @@ function App() {
         <SideWrapper side="right" className="z-50">
           <EditorView />
         </SideWrapper>
-        <Canvas className="z-0 h-full w-full" />
+        <CanvasView />
         <SideWrapper side="left" className="left-4 top-4">
           <Sidebar />
         </SideWrapper>
