+++ apps/frontend/src/components/commons/cursor/index.tsx
