+++ apps/frontend/src/features/canvas/hooks/useCollaborativeCursors.ts
@@ -1,7 +1,8 @@
-import { useReactFlow, type XYPosition } from "@xyflow/react";
-import * as Y from "yjs";
 import { useCallback, useEffect, useRef, useState } from "react";
 import { SocketIOProvider } from "y-socket.io";
+import * as Y from "yjs";
+import { useReactFlow, type XYPosition } from "@xyflow/react";
+
 import useUserStore from "@/store/useUserStore";
 import { createSocketIOProvider } from "@/lib/socketProvider";
