+++ apps/frontend/src/features/canvas/index.ts
@@ -1,0 +1,1 @@
+export { Canvas } from "@/features/canvas/ui/Canvas";
