+++ apps/frontend/src/widgets/CanvasView/index.ts
@@ -1,0 +1,1 @@
+export { CanvasView } from "@/widgets/CanvasView/ui";
